requirements.txt
autopep8==1.5.4
confluent-kafka==1.5.0
kafka-python==2.0.2
pycodestyle==2.6.0
toml==0.10.2
